Updated README.md
I was wrong about just downloading the jar. I don't know what the difference between the jar in the zip and the direct link is at the moment, but it won't run when I download just the jar; even though it works fine when I download the zip, and run it from there.